requirements.txt
amqp==5.1.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==5.0.0
boto3==1.25.2
botocore==1.28.2
bs4==0.0.1
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.22.4
colorama==0.4.4
comtypes==1.1.11
cryptography==38.0.1
cssselect==1.1.0
cx-Freeze==6.10
cx-Logging==3.0
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
dlib==19.23.1
dynaconf==3.1.11
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
face-recognition==1.4.0
face-recognition-models==0.3.0
fake-useragent==0.1.11
fastjsonschema==2.15.3
flake8==5.0.4
Flask==2.1.2
fonttools==4.33.3
hvac==1.0.2
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.3.0
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
kombu==5.2.4
lief==0.12.1
loguru==0.6.0
lxml==4.9.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
mysql-connector-python==8.0.29
nbclient==0.6.2
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
nulltype==2.3.1
numpy==1.22.3
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
openpyxl==3.0.9
packaging==21.3
pandocfilters==1.5.0
parse==1.19.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
plaid==0.1.7
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
psycopg2-binary==2.9.3
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pyee==8.2.2
pyflakes==2.5.0
Pygments==2.12.0
pyhcl==0.4.4
pyparsing==3.0.8
pypiwin32==223
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.18.1
python-dateutil==2.8.2
pyttsx3==2.90
pytz==2022.5
pywin32==304
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.3.0
QtPy==2.1.0
requests==2.27.1
requests-html==0.10.0
s3transfer==0.6.0
scikit-learn==1.0.2
scipy==1.8.0
Send2Trash==1.8.0
shortuuid==1.0.9
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.2.0
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
tzdata==2022.5
urllib3==1.26.9
validators==0.20.0
vine==5.0.0
w3lib==1.22.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
Werkzeug==2.1.2
widgetsnbextension==3.6.0
win32-setctime==1.1.0
worldometer==1.0.1
zipp==3.8.0
